backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-community
langgraph
requests
google-auth
google-auth-oauthlib
google-api-python-client